backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
tavily-python
python-dotenv
pydantic
httpx
